+++ src/components/UI/Home/Home.scss
@@ -31,7 +31,7 @@
 
   .social-group{
     display: flex;
-    margin-top:auto;
+    margin-top: 20rem;
     align-content: space-between;
 
     small{
